+++ src/lib/client/graphql/get_label_id.rs
@@ -13,7 +13,7 @@ const LABEL: &str = "rebase";
 const COLOR: &str = "B22222";
 pub(crate) trait GraphQLGetLabel {
     #[allow(async_fn_in_trait)]
-    async fn get_or_create_label_id(&self) -> Result<String, Error>;
+    async fn get_or_create_label_id(&self, label: Option<&str>) -> Result<String, Error>;
 }
 
 #[derive(Deserialize, Debug, Clone)]
@@ -52,7 +52,7 @@ struct Label {
 
 impl GraphQLGetLabel for Client {
     #[instrument(skip(self))]
-    async fn get_or_create_label_id(&self) -> Result<String, Error> {
+    async fn get_or_create_label_id(&self, label: Option<&str>) -> Result<String, Error> {
         // Get the label ID
         let query = r#"
         query($owner:String!, $name:String!, $label:String!) {
@@ -66,8 +66,11 @@ impl GraphQLGetLabel for Client {
             }
             "#;
 
-        // let label = LABEL.to_string();
-        let label = "test".to_string();
+        let label = if let Some(l) = label {
+            l.to_string()
+        } else {
+            LABEL.to_string()
+        };
 
         let vars = Vars {
             owner: self.owner.clone(),